+++ pkg/sinks/webhooks.go
@@ -38,9 +38,9 @@ func NewWebhookSink(webhookURL string) (*WebhookSink, error) {
 // WriteBatch sends a batch of data to the webhook endpoint
 func (s *WebhookSink) WriteBatch(messages []*utils.CDCMessage) error {
 	for _, message := range messages {
-		decodedMessage, err := message.GetDecodedMessage()
+		decodedMessage, err := buildDecodedMessage(message)
 		if err != nil {
-			return fmt.Errorf("failed to get decoded message: %v", err)
+			return fmt.Errorf("failed to build decoded message: %v", err)
 		}
 
 		jsonData, err := json.Marshal(decodedMessage)
@@ -48,8 +48,7 @@ func (s *WebhookSink) WriteBatch(messages []*utils.CDCMessage) error {
 			return fmt.Errorf("failed to marshal data to JSON: %v", err)
 		}
 
-		err = s.sendWithRetry(jsonData)
-		if err != nil {
+		if err = s.sendWithRetry(jsonData); err != nil {
 			return err
 		}
 	}
