+++ pkg/utils/cdc_message.go
@@ -343,130 +343,7 @@ func EncodeValue(value interface{}, dataType uint32) ([]byte, error) {
 	return ConvertToPgCompatibleOutput(value, dataType)
 }
 
-// GetDecodedColumnValue returns the decoded value of a column
-func (m *CDCMessage) GetDecodedColumnValue(columnName string) (interface{}, error) {
-	colIndex := m.GetColumnIndex(columnName)
-	if colIndex == -1 {
-		return nil, fmt.Errorf("column %s not found", columnName)
-	}
-
-	var data []byte
-	if m.NewTuple != nil && colIndex < len(m.NewTuple.Columns) {
-		data = m.NewTuple.Columns[colIndex].Data
-	} else if m.OldTuple != nil && colIndex < len(m.OldTuple.Columns) {
-		data = m.OldTuple.Columns[colIndex].Data
-	} else {
-		return nil, fmt.Errorf("no data available for column %s", columnName)
-	}
-
-	column := m.Columns[colIndex]
-	return DecodeValue(data, column.DataType)
-}
-
-// GetDecodedMessage returns a map with decoded column values
-func (m *CDCMessage) GetDecodedMessage() (map[string]interface{}, error) {
-	decodedMessage := make(map[string]interface{})
-	decodedMessage["Type"] = m.Type
-	decodedMessage["Schema"] = m.Schema
-	decodedMessage["Table"] = m.Table
-	decodedMessage["ReplicationKey"] = m.ReplicationKey
-	decodedMessage["LSN"] = m.LSN
-	decodedMessage["EmittedAt"] = m.EmittedAt
-
-	if m.NewTuple != nil {
-		newTuple := make(map[string]interface{})
-		for _, col := range m.Columns {
-			value, err := m.GetDecodedColumnValue(col.Name)
-			if err != nil {
-				return nil, fmt.Errorf("failed to decode column %s: %v", col.Name, err)
-			}
-			newTuple[col.Name] = value
-		}
-		decodedMessage["NewTuple"] = newTuple
-	}
-
-	if m.OldTuple != nil {
-		oldTuple := make(map[string]interface{})
-		for _, col := range m.Columns {
-			value, err := m.GetDecodedColumnValue(col.Name)
-			if err != nil {
-				return nil, fmt.Errorf("failed to decode column %s: %v", col.Name, err)
-			}
-			oldTuple[col.Name] = value
-		}
-		decodedMessage["OldTuple"] = oldTuple
-	}
-
-	return decodedMessage, nil
-}
-
 // IsColumnToasted checks if a column was TOASTed
 func (m *CDCMessage) IsColumnToasted(columnName string) bool {
 	return m.ToastedColumns[columnName]
-}
-
-// GetOldColumnValue returns the value from OldTuple
-func (m *CDCMessage) GetOldColumnValue(columnName string) (interface{}, error) {
-	colIndex := m.GetColumnIndex(columnName)
-	if colIndex == -1 {
-		return nil, fmt.Errorf("column %s not found", columnName)
-	}
-
-	if m.OldTuple == nil || colIndex >= len(m.OldTuple.Columns) {
-		return nil, fmt.Errorf("no old value available for column %s", columnName)
-	}
-
-	column := m.Columns[colIndex]
-	data := m.OldTuple.Columns[colIndex].Data
-
-	// Add debug logging
-	decodedValue, err := DecodeValue(data, column.DataType)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode old value for column %s: %v", columnName, err)
-	}
-
-	return decodedValue, nil
-}
-
-// Add helper method to check if a column is part of the primary key
-func (m *CDCMessage) IsPrimaryKeyColumn(columnName string) bool {
-	if m.ReplicationKey.Type != ReplicationKeyPK {
-		return false
-	}
-	for _, col := range m.ReplicationKey.Columns {
-		if col == columnName {
-			return true
-		}
-	}
-	return false
-}
-
-// Add new helper method
-func (m *CDCMessage) IsReplicationKeyColumn(columnName string) bool {
-	if m.ReplicationKey.Type == ReplicationKeyFull {
-		return true // All columns are part of replication key for FULL
-	}
-	for _, col := range m.ReplicationKey.Columns {
-		if col == columnName {
-			return true
-		}
-	}
-	return false
-}
-
-// GetChangedColumns returns columns that have different values between old and new tuples
-func (m *CDCMessage) GetChangedColumns() []string {
-	if m.Type != OperationUpdate || m.OldTuple == nil || m.NewTuple == nil {
-		return nil
-	}
-
-	var changed []string
-	for i, col := range m.Columns {
-		oldVal := m.OldTuple.Columns[i].Data
-		newVal := m.NewTuple.Columns[i].Data
-		if !bytes.Equal(oldVal, newVal) {
-			changed = append(changed, col.Name)
-		}
-	}
-	return changed
 }
